requirements_gradio.txt
gradio==4.44.0
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
presidio-redactor==2.2.33
spacy==3.7.2
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2
groq==0.4.2
ollama==0.1.7
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
Pillow==10.1.0
requests==2.31.0
tiktoken==0.5.2
streamlit==1.29.0
